moved game logic functions to separate file
